refactor: extract protocol.py so client imports without the mcp package
Wire-protocol constants (ports, timeouts, framing) and get_auth_token
lived in helpers.py, which imports mcp.types. That made qgis_mcp.client
unimportable in any environment without the mcp package installed — e.g.
a downstream tool that only wants the socket client, not the MCP server.

Move the stdlib-only constants + auth to a new protocol.py; client
imports from there, helpers re-exports them so server-side callers are
unaffected. No behaviour change.

Authored by Claude (Anthropic AI) in a Claude Code session.
