+++ src/qgis_mcp/helpers.py
@@ -1,48 +1,27 @@
 """Shared helpers for server.py and compound_tools.py.
 
-Imports only from ``mcp`` and stdlib — no circular-import risk.
+Imports only from ``mcp``, stdlib, and the stdlib-only ``protocol``
+module — no circular-import risk. Protocol constants live in
+``protocol.py`` (and are re-exported here) so the client stays
+importable without the ``mcp`` package.
 """
 
 import importlib.metadata
 import json
-import os
-import struct
 
 from mcp.types import Annotations, ImageContent, ResourceLink, TextContent
 
-# ---------------------------------------------------------------------------
-# Protocol constants — single source of truth for defaults across all modules
-# ---------------------------------------------------------------------------
-
-DEFAULT_HOST = "localhost"
-DEFAULT_PORT = 9876
-TIMEOUT_DEFAULT = 30  # seconds — most tool commands
-TIMEOUT_LONG = 60  # seconds — execute_processing, render_map, execute_code, batch
-RECV_CHUNK_SIZE = 65536  # bytes per recv/recv_into call
-MAX_MESSAGE_SIZE = 10 * 1024 * 1024  # 10 MB — plugin-side buffer/message limit
-HEADER_STRUCT = struct.Struct(">I")  # 4-byte big-endian uint32 length prefix
-
-BATCH_BLOCKED_COMMANDS = frozenset(
-    {
-        "execute_code",
-        "remove_layer",
-        "delete_features",
-        "set_setting",
-        "reload_plugin",
-    }
+from qgis_mcp.protocol import (  # noqa: F401 — re-exported for server-side importers
+    BATCH_BLOCKED_COMMANDS,
+    DEFAULT_HOST,
+    DEFAULT_PORT,
+    HEADER_STRUCT,
+    MAX_MESSAGE_SIZE,
+    RECV_CHUNK_SIZE,
+    TIMEOUT_DEFAULT,
+    TIMEOUT_LONG,
+    get_auth_token,
 )
-
-
-def get_auth_token():
-    """Return the shared-secret socket token, or ``None`` when auth is disabled.
-
-    Read from the ``QGIS_MCP_TOKEN`` environment variable. When unset or empty,
-    authentication is off and behaviour is unchanged — the plugin accepts any
-    command (the historical default). When set, the client attaches it to every
-    command and the plugin rejects commands that don't present a matching token.
-    """
-    token = os.environ.get("QGIS_MCP_TOKEN", "").strip()
-    return token or None
 
 
 def enrich_diagnose(result: dict) -> dict:
